util: Switch shortenb/sshortenb to int64
Most calls already happen with `int64` instances, so this change reduces
one type conversion. On 32-bit arch, converting int64 -> int is lossy
and results in incorrect values being print for levels > 2GiB.

+++ leveldb/session_util.go
@@ -24,7 +24,7 @@ type dropper struct {
 
 func (d dropper) Drop(err error) {
 	if e, ok := err.(*journal.ErrCorrupted); ok {
-		d.s.logf("journal@drop %s-%d S·%s %q", d.fd.Type, d.fd.Num, shortenb(e.Size), e.Reason)
+		d.s.logf("journal@drop %s-%d S·%s %q", d.fd.Type, d.fd.Num, shortenb(int64(e.Size)), e.Reason)
 	} else {
 		d.s.logf("journal@drop %s-%d %q", d.fd.Type, d.fd.Num, err)
 	}
